+++ modules/common/guest-platform/src/platform/libraries/std-terminal/compukter/io/StdioSupport.kt
@@ -16,6 +16,8 @@ internal fun stdoutPrint(value: Int): Unit = StdioBindings.write(stdoutInt(value
 
 internal fun stdoutPrint(value: Long): Unit = StdioBindings.write("$value")
 
+internal fun stdoutPrint(value: Float): Unit = StdioBindings.write("$value")
+
 internal fun stdoutPrint(value: Boolean): Unit = StdioBindings.write(if (value) "true" else "false")
 
 internal fun stdoutPrint(value: Char): Unit = StdioBindings.write(stdoutChar(value))
@@ -27,6 +29,8 @@ internal fun stdoutPrintln(value: String): Unit = StdioBindings.write(value + "\
 internal fun stdoutPrintln(value: Int): Unit = StdioBindings.write(stdoutInt(value) + "\n")
 
 internal fun stdoutPrintln(value: Long): Unit = StdioBindings.write("$value\n")
+
+internal fun stdoutPrintln(value: Float): Unit = StdioBindings.write("$value\n")
 
 internal fun stdoutPrintln(value: Boolean): Unit = StdioBindings.write(if (value) "true\n" else "false\n")