+++ modules/common/guest-platform/src/platform/libraries/std-terminal/kotlin/io/Console.kt
@@ -15,6 +15,8 @@ public fun print(value: Int): Unit = stdoutPrint(value)
 
 public fun print(value: Long): Unit = stdoutPrint(value)
 
+public fun print(value: Float): Unit = stdoutPrint(value)
+
 public fun print(value: Boolean): Unit = stdoutPrint(value)
 
 public fun print(value: Char): Unit = stdoutPrint(value)
@@ -26,6 +28,8 @@ public fun println(value: String): Unit = stdoutPrintln(value)
 public fun println(value: Int): Unit = stdoutPrintln(value)
 
 public fun println(value: Long): Unit = stdoutPrintln(value)
+
+public fun println(value: Float): Unit = stdoutPrintln(value)
 
 public fun println(value: Boolean): Unit = stdoutPrintln(value)